+++ src/main/kotlin/server/extension/CoreType.kt
@@ -1,6 +1,0 @@
-package ru.endlesscode.bukkitgradle.server.extension
-
-/** Supported server core types to run. */
-public enum class CoreType {
-    SPIGOT, PAPER
-}