+++ src/test/groovy/ru/endlesscode/bukkitgradle/server/extension/ServerConfigurationSpec.groovy
@@ -19,28 +19,4 @@ class ServerConfigurationSpec extends Specification {
         then:
         ["-Dcom.mojang.eula.agree=true", "-Xmx1G"] == args
     }
-
-    void 'when set existing core - should set core successfully'() {
-        when:
-        serverConfiguration.core = "paper"
-
-        then:
-        CoreType.PAPER == serverConfiguration.coreType
-    }
-
-    void 'when set existing core in mixed case - should set core successfully'() {
-        when:
-        serverConfiguration.core = "Paper"
-
-        then:
-        CoreType.PAPER == serverConfiguration.coreType
-    }
-
-    void 'when set not existing core - should fallback to spigot core'() {
-        when:
-        serverConfiguration.core = "uber-bukkit"
-
-        then:
-        CoreType.SPIGOT == serverConfiguration.coreType
-    }
 }